Combine Base and Runner into a single class
The `Graphiti::Base` module was a legacy from jsonapi_suite that was
mixed into the `Graphiti::Runner` and `Graphiti` modules to provide them
both with functionality.  Turns out we weren't using the latter case
anymore so there's no reason for this to remain a separate concern.
