+++ lib/graphiti/runner.rb
@@ -1,16 +1,23 @@
 module Graphiti
   class Runner
     attr_reader :params
-    include Graphiti::Base
 
     def initialize(resource_class, params, query = nil)
       @resource_class = resource_class
       @params = params
       @query = query
+
+      validator = RequestValidator.new(jsonapi_resource, params)
+
+      @deserialized_payload = validator.deserialized_payload
     end
 
     def jsonapi_resource
       @jsonapi_resource ||= @resource_class.new
+    end
+
+    def deserialized_payload
+      @deserialized_payload
     end
 
     # Typically, this is 'self' of a controller
@@ -21,6 +28,54 @@ module Graphiti
     # end
     def jsonapi_context
       Graphiti.context[:object]
+    end
+
+    def query
+      @query ||= Query.new(jsonapi_resource, params)
+    end
+
+    def query_hash
+      @query_hash ||= query.hash
+    end
+
+    def wrap_context
+      Graphiti.with_context(jsonapi_context, action_name.to_sym) do
+        yield
+      end
+    end
+
+    def jsonapi_context
+      self
+    end
+
+    def jsonapi_scope(scope, opts = {})
+      jsonapi_resource.build_scope(scope, query, opts)
+    end
+
+    def jsonapi_render_options
+      options = {}
+      options.merge!(default_jsonapi_render_options)
+      options[:meta]   ||= {}
+      options[:expose] ||= {}
+      options[:expose][:context] = jsonapi_context
+      options
+    end
+
+    def proxy(base = nil, opts = {})
+      base ||= jsonapi_resource.base_scope
+      scope_opts = opts.slice :sideload_parent_length,
+        :default_paginate,
+        :after_resolve,
+        :sideload,
+        :parent,
+        :params
+      scope = jsonapi_scope(base, scope_opts)
+      ResourceProxy.new jsonapi_resource,
+        scope,
+        query,
+        payload: deserialized_payload,
+        single: opts[:single],
+        raise_on_missing: opts[:raise_on_missing]
     end
   end
 end
